refactor: rename ElementSection/Row/Column to Section/Row/Column, rename GridLeftAndMainExtension
- ElementSection → Section (file, class, table_name)
- ElementRow → Row (file, class, table_name)
- ElementColumn → Column (file, class, table_name)
- ElementalGridLeftAndMainExtension → GridLeftAndMainExtension (file, class)
- Fix stale ElementTreeBuilder → GridTreeBuilder reference in elements.yml
- Update all cross-references in src/ and _config/
